needs to switch to swc from esbuild/jest
