+++ src/plugin.ts
@@ -1,5 +1,7 @@
 import { detect } from 'detect-browser'
 import { PluginEvent, Meta } from '@posthog/plugin-scaffold'
+import { userAgentV3 } from './v3'
+import { userAgentV2 } from './v2'
 
 export type UserAgentPluginConfiguration = {
     enable: string
@@ -18,8 +20,9 @@ export type UserAgentMetaInput = {
     }
 }
 
-interface PluginEventExtra extends PluginEvent {
+export interface PluginEventExtra extends PluginEvent {
     $useragent?: string
+    $raw_user_agent?: string
 }
 
 /**
@@ -86,7 +89,9 @@ export async function processEvent(event: PluginEventExtra, { global }: Meta<Use
         delete event.properties.$user_agent
     }
 
-    if (!userAgent || userAgent === '') {
+    const hasRawUserAgentKey = availableKeysOfEvent.includes('$raw_user_agent')
+    const hasUserAgent = userAgent.length > 0
+    if (!hasRawUserAgentKey && !hasUserAgent) {
         if (global.debugMode) {
             console.warn(`UserAgentPlugin.processEvent(): $useragent is empty`)
         }
@@ -94,9 +99,9 @@ export async function processEvent(event: PluginEventExtra, { global }: Meta<Use
         return event
     }
 
-    const agentInfo = detect(userAgent)
-    const device = detectDevice(userAgent)
-    const deviceType = detectDeviceType(userAgent)
+    if (hasRawUserAgentKey && !hasUserAgent) {
+        return userAgentV3(event, userAgent, global)
+    }
 
     const eventProperties = Object.keys(event.properties)
     const hasBrowserProperties = eventProperties.some((value: string) =>
@@ -113,50 +118,5 @@ export async function processEvent(event: PluginEventExtra, { global }: Meta<Use
         return event
     }
 
-    // The special Posthog property names are retrieved from:
-    // https://github.com/PostHog/posthog/blob/master/frontend/src/lib/components/PropertyKeyInfo.tsx
-    event.properties['$device'] = device
-    event.properties['$device_type'] = deviceType
-
-    if (agentInfo) {
-        event.properties['$browser'] = agentInfo.name
-        event.properties['$browser_version'] = agentInfo.version
-        event.properties['$os'] = agentInfo.os
-        // Custom property
-        event.properties['$browser_type'] = agentInfo.type
-    }
-
-    return event
-}
-
-// detectDevice and detectDeviceType from https://github.com/PostHog/posthog-js/blob/9abedce5ac877caeb09205c4b693988fc09a63ca/src/utils.js#L808-L837
-function detectDevice(userAgent: string) {
-    if (/Windows Phone/i.test(userAgent) || /WPDesktop/.test(userAgent)) {
-        return 'Windows Phone'
-    } else if (/iPad/.test(userAgent)) {
-        return 'iPad'
-    } else if (/iPod/.test(userAgent)) {
-        return 'iPod Touch'
-    } else if (/iPhone/.test(userAgent)) {
-        return 'iPhone'
-    } else if (/(BlackBerry|PlayBook|BB10)/i.test(userAgent)) {
-        return 'BlackBerry'
-    } else if (/Android/.test(userAgent) && !/Mobile/.test(userAgent)) {
-        return 'Android Tablet'
-    } else if (/Android/.test(userAgent)) {
-        return 'Android'
-    } else {
-        return ''
-    }
-}
-
-function detectDeviceType(userAgent: string) {
-    const device = detectDevice(userAgent)
-    if (device === 'iPad' || device === 'Android Tablet') {
-        return 'Tablet'
-    } else if (device) {
-        return 'Mobile'
-    } else {
-        return 'Desktop'
-    }
+    return userAgentV2(event, userAgent, global)
 }
